fix int size, small bump
FossilOrigin-Name: e3f7eeb26a2ec1a593657555cdfd20d7b24633ac018249f645540467951b9dec

+++ src/sql.rs
@@ -19,7 +19,7 @@ use sqlx::{
 
 #[derive(sqlx::FromRow, Debug)]
 pub struct List {
-	pub source_id: i64,
+	pub source_id: i32,
 	pub channel: String,
 	pub enabled: bool,
 	pub url: String,
